+++ src/test-misc/csharp-spec2/Spec2Sm.cs
@@ -391,9 +391,6 @@ namespace Csharp.Spec2smTests
                 // Step 1: execute action `trace("Exit Spec2Sm.");`
                 MainClass.Trace("Exit Spec2Sm.");
             } // end of behavior for ROOT
-
-            // State machine root is a special case. It cannot be exited. Mark as unused.
-            _ = this;
         }